splitting: make records private
